requirements.txt
Flask==3.1.1
flask-cors==6.0.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
PyJWT==2.10.1
Flask-Caching==2.3.0
redis==5.0.8
sqlalchemy==2.0.23
clerk-backend-api==3.3.0

APScheduler==3.10.1
